Implement support for internal domain name
