+++ c2c-extension-test/src/test/resources/service/internal-domain/Ballerina.toml
@@ -1,0 +1,8 @@
+[package]
+org = "wso2"
+name = "internal"
+version = "0.1.0"
+
+[build-options]
+observabilityIncluded = true
+cloud = "k8s"
