+++ c2c-extension-test/src/test/resources/service/internal-domain/Cloud.toml
@@ -1,0 +1,2 @@
+[cloud.deployment]
+internal_domain_name="hello-svc"
