+++ c2c-extension-test/src/test/resources/service/invalid-internal-domain/Cloud.toml
@@ -1,0 +1,7 @@
+[container.image]
+repository= "anuruddhal" # optional default is local
+name="hello-api" # optional
+tag="sample2"  # default is latest
+
+[cloud.deployment]
+internal_domain_name="hello_svc"
