+++ c2c-extension-test/src/test/resources/testng-integration.xml
@@ -46,6 +46,7 @@
             <class name="io.ballerina.c2c.test.codeactions.CloudProbesTest"/>
             <class name="io.ballerina.c2c.test.completion.CloudTest"/>
             <class name="io.ballerina.c2c.test.docker.CmdTest"/>
+            <class name="io.ballerina.c2c.test.ServiceTest"/>
         </classes>
     </test>
