+++ c2c-extension/src/main/java/io/ballerina/c2c/models/DeploymentModel.java
@@ -70,6 +70,7 @@ public class DeploymentModel extends KubernetesModel {
     private boolean uberJar;
     private String dockerConfigPath;
     private ResourceRequirements resourceRequirements;
+    private String internalDomainName;
 
     public DeploymentModel() {
         // Initialize with default values.
