+++ c2c-extension/src/main/resources/c2c-schema.json
@@ -146,7 +146,7 @@
           "properties": {
             "internal_domain_name": {
               "type": "string",
-              "pattern": "^(?!\\s*$).+"
+              "pattern": "[a-z0-9]([-a-z0-9]*[a-z0-9])?"
             },
             "external_accessible": {
               "type": "boolean"
